# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DocBook API",
    "version": "1.0.0",
    "description": "Multi-tenant healthcare booking platform.\n\nGenerated from the route handlers and their Zod validation schemas\n(`npx tsx scripts/generate-openapi.ts`) — not hand-written, so it cannot drift from\nwhat the code actually validates. Generation fails if a handler has no registry entry.\n\nResponse bodies are described by status code and the standard envelope rather than by\nper-field schemas: handlers return Prisma rows, no response Zod schemas exist, and\ninventing approximations would produce a document that lies in a way nobody can check."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
...
  "paths": {
...
    "/api/v1/account/password": {
      "post": {
        "summary": "Change the signed-in account's own password. Always requires the current password, and clears the mustChangePassword flag.",
        "description": "Requires an authenticated session; the handler performs its own ownership check.",
        "tags": [
          "Account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied. An audit row is written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie. Obtained via POST /api/auth/callback/credentials. No route trusts a client-supplied tenantId or role — both come from the verified session."
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "data": {},
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "data"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "SLOT_TAKEN"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}